requirements.txt
asgiref==3.6.0
certifi==2022.12.7
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.9.0
pbr==5.11.1
Pillow==10.0.0
pipenv==2022.12.19
platformdirs==2.6.2
psycopg2==2.9.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.3
stevedore==5.0.0
tzdata==2022.7
virtualenv==20.17.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
whitenoise==6.6.0
